Varianta noua orar

+++ app/src/main/java/com/example/licenta/BottomNavigationView/FragmentNote/NoteFragment.java
@@ -40,7 +40,6 @@ import java.util.Map;
 public class NoteFragment extends Fragment implements RecyclerViewAdapterMaterii.OnButtonClickListener {
 
     ArrayList<Materie> materii = new ArrayList<Materie>();
-    List<Float> listaNote = new ArrayList<Float>();
     ArrayList<Float> listaNoteINDIVIDUALA;
 
     RecyclerViewAdapterMaterii adapterMaterii;
@@ -75,27 +74,26 @@ public class NoteFragment extends Fragment implements RecyclerViewAdapterMaterii
         String collectionName = "NoteMaterii";
 
         DocumentReference documentReference = db.collection("users").document(userID);
-        CollectionReference colectie=documentReference.collection(collectionName);
+        CollectionReference colectie = documentReference.collection(collectionName);
 
-        if(materii.size()==0){
+        if (materii.size() == 0) {
             colectie.get().addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
-                        @Override
-                        public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
-                            for(QueryDocumentSnapshot queryDocumentSnapshots1: queryDocumentSnapshots){
-                                Materie materie=queryDocumentSnapshots1.toObject(Materie.class);
-                                String materieID= queryDocumentSnapshots1.getId();
-                                materie.setId(materieID);
-                                materii.add(materie);
-                                adapterMaterii.notifyDataSetChanged();
-                            }
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Log.d("NOTOK", e.getMessage().toString());
-                        }
-                    });
+                @Override
+                public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
+                    for (QueryDocumentSnapshot queryDocumentSnapshots1 : queryDocumentSnapshots) {
+                        Materie materie = queryDocumentSnapshots1.toObject(Materie.class);
+                        String materieID = queryDocumentSnapshots1.getId();
+                        materie.setId(materieID);
+                        materii.add(materie);
+                        adapterMaterii.notifyDataSetChanged();
+                    }
+                }
+            }).addOnFailureListener(new OnFailureListener() {
+                @Override
+                public void onFailure(@NonNull Exception e) {
+                    Log.d("NOTOK", e.getMessage().toString());
+                }
+            });
 
         }
 
@@ -123,30 +121,28 @@ public class NoteFragment extends Fragment implements RecyclerViewAdapterMaterii
                 if (creareListaNote()) {
 
                     if (!edtNameMaterie.getText().toString().equals("")) {
-                         materie = new Materie(edtNameMaterie.getText().toString(), listaNoteINDIVIDUALA);
+                        materie = new Materie(edtNameMaterie.getText().toString(), listaNoteINDIVIDUALA);
                         materii.add(materie);
                         adapterMaterii.notifyItemInserted(materii.indexOf(materie));
 
-                        Log.d("mesaj",colectie.getPath());
+                        Log.d("mesaj", colectie.getPath());
 
                         Map<String, Object> NoteMaterii = new HashMap<>();
-                        NoteMaterii.put("denumire",materie.getDenumire());
+                        NoteMaterii.put("denumire", materie.getDenumire());
                         NoteMaterii.put("listanote", listaNoteINDIVIDUALA);
 
 
-                        colectie.add(NoteMaterii)
-                                .addOnSuccessListener(new OnSuccessListener<DocumentReference>() {
-                                    @Override
-                                    public void onSuccess(DocumentReference documentReference) {
-                                        Log.d("OK",documentReference.getPath().toString());
-                                    }
-                                })
-                                .addOnFailureListener(new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        Log.d("NOTOK",e.getMessage().toString());
-                                    }
-                                });
+                        colectie.add(NoteMaterii).addOnSuccessListener(new OnSuccessListener<DocumentReference>() {
+                            @Override
+                            public void onSuccess(DocumentReference documentReference) {
+                                Log.d("OK", documentReference.getPath().toString());
+                            }
+                        }).addOnFailureListener(new OnFailureListener() {
+                            @Override
+                            public void onFailure(@NonNull Exception e) {
+                                Log.d("NOTOK", e.getMessage().toString());
+                            }
+                        });
 
 //                Map<String ,Object> user= new HashMap<>();
 //                user.put("notePeMaterie",materiePreluata.getDenumire());
@@ -168,12 +164,13 @@ public class NoteFragment extends Fragment implements RecyclerViewAdapterMaterii
 
             dialog.show();
         });
+
         return view;
     }
 
     private boolean creareListaNote() {
         listaNoteINDIVIDUALA = new ArrayList<>();
-//        materii.clear();
+
         boolean result = true;
         for (int i = 0; i < layout_listaNote.getChildCount(); i++) {
             View notaview = layout_listaNote.getChildAt(i);
@@ -181,7 +178,6 @@ public class NoteFragment extends Fragment implements RecyclerViewAdapterMaterii
 
             if (!editTextNota.getText().toString().equals("")) {
                 listaNoteINDIVIDUALA.add(Float.valueOf(editTextNota.getText().toString()));
-//                listaNote.add(Float.valueOf(editTextNota.getText().toString()));
 
 
             } else {
@@ -275,18 +271,18 @@ public class NoteFragment extends Fragment implements RecyclerViewAdapterMaterii
 
         btn_saveNote.setOnClickListener(view2 -> {
 
-//            if (creareListaNote()) {
+            if (creareListaNote()) {
 
-            if (!edtNameMaterie.getText().toString().equals("")) {
-                Materie materie = new Materie(edtNameMaterie.getText().toString(), listaNoteINDIVIDUALA);
-                materii.set(index, materie);
-                adapterMaterii.notifyItemChanged(index);
+                if (!edtNameMaterie.getText().toString().equals("")) {
+                    Materie materie = new Materie(edtNameMaterie.getText().toString(), listaNoteINDIVIDUALA);
+                    materii.set(index, materie);
+                    adapterMaterii.notifyItemChanged(index);
+                } else {
+                    Toast.makeText(getContext(), "Nume materie lipsa!", Toast.LENGTH_SHORT).show();
+                }
             } else {
-                Toast.makeText(getContext(), "Nume materie lipsa!", Toast.LENGTH_SHORT).show();
+                Toast.makeText(getContext(), "Introdu nota!", Toast.LENGTH_SHORT).show();
             }
-//            } else {
-//                Toast.makeText(getContext(), "Introdu nota!", Toast.LENGTH_SHORT).show();
-//            }
             dialog.dismiss();
         });
 
@@ -297,45 +293,42 @@ public class NoteFragment extends Fragment implements RecyclerViewAdapterMaterii
 
     @Override
     public void onDeleteClick(int position) {
-//-------------------------face stergere in firestore si pe ecran, insa da eroare cand positie==materii.size()
-
-//        if(position < materii.size()){
-//            materii.remove(position);
-//            adapterMaterii.notifyItemRemoved(position);
-//        }
-
         if (materii.size() > 0) {
-            Materie materie = materii.get(position);
-            materii.remove(materie);
-            adapterMaterii.notifyItemRemoved(materii.indexOf(materie));
 
-            //stergere in firestore a materiei din coletie
-            //functie care imi preia notele din BD initilizez instanta FIRESTORE
-            FirebaseFirestore db = FirebaseFirestore.getInstance();
-            FirebaseAuth mAuth = FirebaseAuth.getInstance();
-            String userID = mAuth.getCurrentUser().getUid();
-            String collectionName = "NoteMaterii";
+            if (position >= 0 && position < materii.size()) {
+                Materie materie = materii.get(position);
+                adapterMaterii.notifyItemRemoved(materii.indexOf(materie));
+                materii.remove(materie);
 
-            DocumentReference documentReference = db.collection("users").document(userID);
-            CollectionReference colectie=documentReference.collection(collectionName);
+                System.out.println(position);
 
-            DocumentReference elemSters=colectie.document(materie.getId());
-            elemSters.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                @Override
-                public void onSuccess(Void unused) {
-                    Log.d(TAG, "DocumentSnapshot successfully deleted!");
-                }
-            }).addOnFailureListener(new OnFailureListener() {
-                @Override
-                public void onFailure(@NonNull Exception e) {
-                    Log.w(TAG, "Error deleting document", e);
-                }
-            });
+
+                //stergere in firestore a materiei din coletie
+                //functie care imi preia notele din BD initilizez instanta FIRESTORE
+                FirebaseFirestore db = FirebaseFirestore.getInstance();
+                FirebaseAuth mAuth = FirebaseAuth.getInstance();
+                String userID = mAuth.getCurrentUser().getUid();
+                String collectionName = "NoteMaterii";
+
+                DocumentReference documentReference = db.collection("users").document(userID);
+                CollectionReference colectie = documentReference.collection(collectionName);
+
+                DocumentReference elemSters = colectie.document(materie.getId());
+                elemSters.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                    @Override
+                    public void onSuccess(Void unused) {
+                        Log.d(TAG, "DocumentSnapshot successfully deleted!");
+                    }
+                }).addOnFailureListener(new OnFailureListener() {
+                    @Override
+                    public void onFailure(@NonNull Exception e) {
+                        Log.w(TAG, "Error deleting document", e);
+                    }
+                });
+
+            }
 
         }
 
-//        Materie materie=materii.get(position);
-//        materii.remove(materie);
-//        adapterMaterii.notifyItemRemoved(position);
     }
 }
